Add models file
